Add files via upload
This project seeks to analysis the possible factors that influence the presence of cardiovascular diseases. To perform this analysis, dataset was derived from an online source (kaggle.com) and it covers approximately 70,000 patients’ records with the response variable as presence or absence of cardiovascular disease, this is a categorical variable (1 = present and 0 = absent) was used. We developed four models in our analysis: decision tree, KNN, random forest, and Gaussian Naive Bayes. The results show that the decision tree model had the highest accuracy. Furthermore, the decision tree analysis revealed that the most important factors associated with vascular diseases are systolic blood pressure, age, and cholesterol. Therefore, it is strongly advised to monitor and control these significant risk factors in order to reduce the risk of developing vascular diseases. The description of the variables can be found here :  https://www.kaggle.com/sulianova/cardiovascular